+++ src/grbl/serial.h
@@ -55,10 +55,10 @@ uint8_t serial_read();
 void serial_reset_read_buffer();
 
 // Returns the number of bytes used in the RX serial buffer.
-uint8_t serial_get_rx_buffer_count();
+uint16_t serial_get_rx_buffer_count();
 
 // Returns the number of bytes used in the TX serial buffer.
 // NOTE: Not used except for debugging and ensuring no TX bottlenecks.
-uint8_t serial_get_tx_buffer_count();
+uint16_t serial_get_tx_buffer_count();
 
 #endif
